New products.ts file and added empty LS logic
