+++ src/index.html
@@ -32,7 +32,7 @@
               id="products"
               class="menu__list__link__anchor"
               href="./html/products.html"
-              >Alla Produkter</a
+              >Produkter</a
             >
           </li>
           <li class="menu__list__link">
@@ -61,7 +61,10 @@
           Upptäck de senaste mobilskalen till de bästa priserna
         </h2>
 
-        <a class="hero__container__link" href="./html/products.html">
+        <a
+          id="all-products-link-button"
+          class="hero__container__link"
+          href="./html/products.html">
           <button class="hero__container__link__btn">Handla</button>
         </a>
       </div>
